+++ src/kern/ipc.c
@@ -2,8 +2,8 @@
 #include <kern/ipc.h>
 #include <kern/tasks.h>
 
-static int msg_copy(Message *dest, Message *source) {
-    unsigned int size = source->len < dest->len ? source->len : dest->len;
+static size_t msg_copy(Message *dest, Message *source) {
+    size_t size = source->len < dest->len ? source->len : dest->len;
     memcpy(dest->array, source->array, size);
     return size;
 }
@@ -18,11 +18,11 @@ static void ipc_recvsend(Task *receiver, Task *sender) {
     sender->status = REPLYBLOCKED;
 }
 
-void ipc_send(int tid, int recvtid, char *msg, int msglen, char *reply, int rplen) {
+void ipc_send(int tid, int recvtid, char *msg, size_t msglen, char *reply, size_t rplen) {
     Task *current_task = task_at(tid);
     Task *recv_task = task_at(recvtid);
     if (!recv_task || !ipc_connectable(recv_task)) {
-        current_task->tf->r0 = -1;
+        current_task->tf->r0 = (uint32_t) -1;
         return;
     }
     current_task->send_msg.array = msg;
@@ -39,7 +39,7 @@ void ipc_send(int tid, int recvtid, char *msg, int msglen, char *reply, int rple
     }
 }
 
-void ipc_receive(int tid, int *sendtid, char *msg, int msglen) {
+void ipc_receive(int tid, int *sendtid, char *msg, size_t msglen) {
     Task *current_task = task_at(tid);
 
     current_task->send_tid = sendtid;
@@ -54,22 +54,22 @@ void ipc_receive(int tid, int *sendtid, char *msg, int msglen) {
     }
 }
 
-void ipc_reply(int tid, int replytid, char *reply, int rplen) {
+void ipc_reply(int tid, int replytid, char *reply, size_t rplen) {
     Task *current_task = task_at(tid);
     Task *reply_task = task_at(replytid);
     if (!reply_task || !ipc_connectable(reply_task)) {
-        current_task->tf->r0 = -1;
+        current_task->tf->r0 = (uint32_t) -1;
         return;
     }
     if (reply_task->status != REPLYBLOCKED) {
-        current_task->tf->r0 = -2;
+        current_task->tf->r0 = (uint32_t) -2;
         return;
     }
     Message msg = {
         .array = reply,
         .len = rplen
     };
-    int length = msg_copy(&reply_task->reply_msg, &msg);
+    size_t length = msg_copy(&reply_task->reply_msg, &msg);
     reply_task->tf->r0 = length;
     current_task->tf->r0 = length;
     reply_task->status = READY;
@@ -79,7 +79,7 @@ void ipc_cleanup(int tid) {
     Task *current_task = task_at(tid);
     while (queue_size(&current_task->send_queue)) {
         Task *task = task_at(queue_pop(&current_task->send_queue));
-        task->tf->r0 = -2;
+        task->tf->r0 = (uint32_t) -2;
         task->status = READY;
     }
 }
